backend/requirements.txt
numpy==1.26.0
torch
torchvision
flask
pillow
